Moved aristotle_validate in a separate module
